+++ src/main/java/nextstep/qna/domain/answer/Answer.java
@@ -4,8 +4,8 @@ import nextstep.qna.CannotDeleteException;
 import nextstep.qna.NotFoundException;
 import nextstep.qna.UnAuthorizedException;
 import nextstep.qna.domain.ContentType;
-import nextstep.qna.domain.DeleteHistory;
-import nextstep.qna.domain.Question;
+import nextstep.qna.domain.deletehistory.DeleteHistory;
+import nextstep.qna.domain.question.Question;
 import nextstep.users.domain.NsUser;
 
 import java.time.LocalDateTime;
