+++ src/main/java/nextstep/qna/domain/answer/Answers.java
@@ -1,8 +1,8 @@
 package nextstep.qna.domain.answer;
 
 import nextstep.qna.CannotDeleteException;
-import nextstep.qna.domain.DeleteHistory;
-import nextstep.qna.domain.Question;
+import nextstep.qna.domain.deletehistory.DeleteHistory;
+import nextstep.qna.domain.question.Question;
 import nextstep.users.domain.NsUser;
 
 import java.time.LocalDateTime;
