+++ src/main/java/nextstep/qna/domain/question/Question.java
@@ -1,8 +1,10 @@
-package nextstep.qna.domain;
+package nextstep.qna.domain.question;
 
 import nextstep.qna.CannotDeleteException;
+import nextstep.qna.domain.ContentType;
 import nextstep.qna.domain.answer.Answer;
 import nextstep.qna.domain.answer.Answers;
+import nextstep.qna.domain.deletehistory.DeleteHistory;
 import nextstep.users.domain.NsUser;
 
 import java.time.LocalDateTime;
